% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minigr_chm13}
\alias{minigr_chm13}
\title{GRanges representation of minigraph for HPRC v1.0}
\format{
GRanges including DNAStringSets for bubble sequence, CHMv13 version
}
\usage{
minigr_chm13
}
\description{
GRanges representation of minigraph for HPRC v1.0
}
\keyword{datasets}
