% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minigr_GRCh38}
\alias{minigr_GRCh38}
\title{GRanges representation of minigraph for HPRC v1.0}
\format{
GRanges including DNAStringSets for bubble sequence, GRCh38 version
}
\usage{
minigr_GRCh38
}
\description{
GRanges representation of minigraph for HPRC v1.0
}
\keyword{datasets}
