% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePCs.R
\name{simplePCs}
\alias{simplePCs}
\title{produce colored display of PCA of simple (putatively biallelic) genotypes from HPRC}
\usage{
simplePCs(clvcf, particdf, pop = "super")
}
\arguments{
\item{clvcf}{CollapsedVCF instance from VariantAnnotation}

\item{particdf}{data.frame or tibble denoting populations etc. of HPRC participants}

\item{pop}{character(1) "super" or "sub" to denote population type used to color points}
}
\description{
produce colored display of PCA of simple (putatively biallelic) genotypes from HPRC
}
\examples{
data("smyd3v", package="BiocHPRClk")
data(hprc_partic)
simplePCs(smyd3v, hprc_partic)
}
