% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf1Mb}
\alias{vcf1Mb}
\title{small excerpt from chr1 vcf, GRCh38 coordinates}
\format{
CollapsedVCF instance (VariantAnnotation)
}
\usage{
vcf1Mb
}
\description{
small excerpt from chr1 vcf, GRCh38 coordinates
}
\keyword{datasets}
